Introduce OpenShift template with restart policy set to Never
This new template is for development purposes only. It deploys the
project as a Pod so restart policy can be set to Never. Standard
deployment configs only allow "Always" restart policy.

There is also `template-prod.yaml` which is intended for
staging/production deployments. saas-herder automatically sets restart
policy to "Always" for those environments.